requirements.txt
Flask
Flask-Assets
cssmin
jsmin
Flask-PyMongo
python-dotenv
opencv-python